pyproject.toml
[tool.poetry]
name = "fashion-datalake"
version = "0.1.0"
description = "Create and maintain a continously updating hive datalake."
authors = ["Giulia Caglia"]

[tool.poetry.dependencies]
python = "^3.9"
Scrapy = "^2.6.1"
scrapy-splash = "^0.8.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pycodestyle = "^2.8.0"
pydocstyle = "^6.1.1"
pylama = "^8.3.8"
pep8 = "^1.7.1"
autopep8 = "^1.6.0"
black = "^22.3.0"
mypy = "^0.942"
pylint = "^2.13.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
